refactor conv_block
