+++ dataset/models/tf/layers/resize.py
@@ -2,6 +2,7 @@
 import numpy as np
 import tensorflow as tf
 
+from .block import _conv_block as conv_block
 from .conv import conv, conv1d_transpose_nn
 from .core import xip
 
@@ -105,7 +106,7 @@ def _depth_to_space(inputs, block_size, name='d2s'):
     return x
 
 
-def subpixel_conv(inputs, factor=2, name=None, data_format='channels_last', **kwargs):
+def subpixel_conv(inputs, factor=2, name='subpixel', data_format='channels_last', **kwargs):
     """ Resize input tensor with subpixel convolution (depth to space operation)
 
     Parameters
@@ -124,20 +125,16 @@ def subpixel_conv(inputs, factor=2, name=None, data_format='channels_last', **kw
     tf.Tensor
     """
     dim = inputs.shape.ndims - 2
-    if dim == 3:
-        dafo = 'NDHWC' if data_format == 'channels_last' else 'NCDHW'
-    else:
-        dafo = 'NHWC' if data_format == 'channels_last' else 'NCHW'
 
     _, channels = _calc_size(inputs, factor, data_format)
 
     with tf.variable_scope(name):
-        x = conv(inputs, filters=channels*factor**dim, kernel_size=1, name='conv', **kwargs)
-        x = tf.depth_to_space(x, block_size=factor, name='d2s', data_format=dafo)
+        x = conv(inputs, filters=channels*factor**dim, kernel_size=1, name='conv', data_format = data_format, **kwargs)
+        x = depth_to_space(x, block_size=factor, name='d2s', data_format=data_format)
     return x
 
 
-def resize_bilinear_additive(inputs, factor=2, name=None, data_format='channels_last', **kwargs):
+def resize_bilinear_additive(inputs, factor=2, name='bilinear_additive', data_format='channels_last', **kwargs):
     """ Resize input tensor with bilinear additive technique
 
     Parameters
@@ -156,16 +153,11 @@ def resize_bilinear_additive(inputs, factor=2, name=None, data_format='channels_
     tf.Tensor
     """
     dim = inputs.shape.ndims - 2
-    size, channels = _calc_size(inputs, factor, data_format)
+    _, channels = _calc_size(inputs, factor, data_format)
     layout = kwargs.get('layout', 'c')
     with tf.variable_scope(name):
-        x = inputs
-        if data_format == 'channels_first':
-            x = tf.transpose(x, [0, 2, 3, 1])
-        x = tf.image.resize_bilinear(x, size=size, name='resize')
-        if data_format == 'channels_first':
-            x = tf.transpose(x, [0, 3, 1, 2])
-        x = conv(x, layout, filters=channels*factor**dim, kernel_size=1, name='conv', **kwargs)
+        x = resize_bilinear(inputs, factor, name=name, data_format=data_format, **kwargs)
+        x = conv_block(x, layout, filters=channels*factor**dim, kernel_size=1, name='conv', **kwargs)
         x = xip(x, depth=factor**dim, reduction='sum', name='addition')
     return x
